Replace BackgroundImage with BackgroundMedia
Introduce BackgroundMedia component that supports both image and video backgrounds (videoSrc or image src) and migrate usages across the app. Removed the old BackgroundImage component and updated stories/tests/refs to BackgroundMedia. Added hero_background.mp4 to public/videos for video backgrounds and updated two image assets and an image alt text for the history card. Also cleaned up a small comment block in IconYoutube.
