+++ src/components/Hero/Hero.tsx
@@ -1,15 +1,12 @@
-import { BackgroundImage } from "../shared/BackgroundImage/BackgroundImage";
+import { BackgroundMedia } from "../shared/BackgroundMedia/BackgroundMedia";
 import Container from "../shared/Container";
 import { Heading } from "../shared/Heading/Heading";
 import LogoAK from "@/icons/LogoAK";
 
 export const Hero = () => (
   <section className="relative h-[597px] overflow-hidden tablet:h-[614px] desktop:h-[651px]">
-    <BackgroundImage
-      alt="background image"
-      src="/images/hero_background.webp"
-      priority
-      quality={100}
+    <BackgroundMedia
+      videoSrc="/videos/hero_background.mp4"
       className="object-center contrast:hidden tablet:object-[80%]"
     >
       <Container className="mx-auto flex h-full flex-col items-center justify-center gap-6 pb-5 pt-20 text-center text-white tablet:pt-32 desktop:pt-28">
@@ -32,6 +29,6 @@ export const Hero = () => (
         </Heading>
         <LogoAK className="mt-2 h-[32px] w-[88px] opacity-[33%] tablet:mt-44 desktop:mt-48" />
       </Container>
-    </BackgroundImage>
+    </BackgroundMedia>
   </section>
 );
